backend/requirements.txt
# FasalVaidya Backend Dependencies
# =================================
# Python 3.11+ required

# Flask Web Framework
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# Supabase (for Storage and Database)
supabase>=2.3.0

# Database
# SQLite is built-in, no additional package needed

# ML/AI Dependencies
tensorflow>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0

# ML Training (required for model training)
scikit-learn>=1.3.0
matplotlib>=3.8.0
tqdm>=4.66.0

# Image Processing
imageio>=2.31.0
opencv-python>=4.8.0  # Optional: for Grad-CAM heatmaps

# Report Generation & Export
reportlab>=4.0.0
openpyxl>=3.1.0

# Utilities
python-dateutil>=2.8.0
# Testing
pytest>=7.4.0
